requirements.txt
# Requirements for ABAplay Email Automation
streamlit>=1.31.0
resend>=0.8.0
dnspython>=2.4.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
